# More efficient float(T) for rationals and big*, works with Irrational