Before:
```julia
julia> T = BigInt; @benchmark float($T)
BenchmarkTools.Trial: 
  memory estimate:  160 bytes
  allocs estimate:  4
  --------------
  minimum time:     171.846 ns (0.00% GC)
  median time:      178.750 ns (0.00% GC)
  mean time:        250.258 ns (27.08% GC)
  maximum time:     12.712 μs (97.33% GC)
  --------------
  samples:          10000
  evals/sample:     800

julia> T = BigFloat; @benchmark float($T)
BenchmarkTools.Trial: 
  memory estimate:  104 bytes
  allocs estimate:  2
  --------------
  minimum time:     109.191 ns (0.00% GC)
  median time:      117.609 ns (0.00% GC)
  mean time:        156.539 ns (22.95% GC)
  maximum time:     8.991 μs (97.65% GC)
  --------------
  samples:          10000
  evals/sample:     934

julia> T = Rational{Int64}; @benchmark float($T)
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     70.327 ns (0.00% GC)
  median time:      71.700 ns (0.00% GC)
  mean time:        72.825 ns (0.00% GC)
  maximum time:     125.440 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     975

julia> T = Rational{BigInt}; @benchmark float($T)
BenchmarkTools.Trial: 
  memory estimate:  624 bytes
  allocs estimate:  17
  --------------
  minimum time:     511.825 ns (0.00% GC)
  median time:      548.327 ns (0.00% GC)
  mean time:        843.876 ns (33.19% GC)
  maximum time:     55.189 μs (98.20% GC)
  --------------
  samples:          10000
  evals/sample:     194
```
after:
```julia
julia> T = BigInt; @benchmark float($T)
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     0.013 ns (0.00% GC)
  median time:      0.020 ns (0.00% GC)
  mean time:        0.021 ns (0.00% GC)
  maximum time:     0.047 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1000

julia> T = BigFloat; @benchmark float($T)
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     0.013 ns (0.00% GC)
  median time:      0.021 ns (0.00% GC)
  mean time:        0.021 ns (0.00% GC)
  maximum time:     0.042 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1000

julia> T = Rational{Int64}; @benchmark float($T)
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     0.013 ns (0.00% GC)
  median time:      0.020 ns (0.00% GC)
  mean time:        0.021 ns (0.00% GC)
  maximum time:     0.174 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1000

julia> T = Rational{BigInt}; @benchmark float($T)
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     0.013 ns (0.00% GC)
  median time:      0.020 ns (0.00% GC)
  mean time:        0.020 ns (0.00% GC)
  maximum time:     0.040 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1000
```