@fredrikekre 2017-03-29
Note https://github.com/JuliaCI/BenchmarkTools.jl/issues/50 for interpolating types. (PR is still much better than master though)