@giordano 2017-03-29
Ok, maybe these benchmarks are more reliable (only before the PR, after the result is just the same):
```julia
julia> f() = float(BigInt); @benchmark f()
BenchmarkTools.Trial: 
  memory estimate:  160 bytes
  allocs estimate:  4
  --------------
  minimum time:     104.704 ns (0.00% GC)
  median time:      111.595 ns (0.00% GC)
  mean time:        186.148 ns (38.13% GC)
  maximum time:     12.549 μs (98.30% GC)
  --------------
  samples:          10000
  evals/sample:     943

julia> f() = float(BigFloat); @benchmark f()
BenchmarkTools.Trial: 
  memory estimate:  104 bytes
  allocs estimate:  2
  --------------
  minimum time:     53.766 ns (0.00% GC)
  median time:      64.738 ns (0.00% GC)
  mean time:        103.334 ns (36.33% GC)
  maximum time:     9.378 μs (98.56% GC)
  --------------
  samples:          10000
  evals/sample:     984

julia> f() = float(Rational{Int64}); @benchmark f()
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     23.922 ns (0.00% GC)
  median time:      24.036 ns (0.00% GC)
  mean time:        24.206 ns (0.00% GC)
  maximum time:     46.649 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     996

julia> f() = float(Rational{BigInt}); @benchmark f()
BenchmarkTools.Trial: 
  memory estimate:  624 bytes
  allocs estimate:  17
  --------------
  minimum time:     443.935 ns (0.00% GC)
  median time:      482.889 ns (0.00% GC)
  mean time:        775.923 ns (36.09% GC)
  maximum time:     58.559 μs (97.61% GC)
  --------------
  samples:          10000
  evals/sample:     199
```