@ararslan 2017-04-04
> Is this good to be merged?

Yes

> Should wait for the 0.6 branching (the patch introduces a new method)?

Probably