@giordano 2017-05-02
Patch rebased on master ;-)  I also added methods for `Complex{T}`.